kerberos_auth: make it possible to set the kerberos hostname via the environment
Using `getfqdn` does not work well in containerized environments, as it
would return the pod name (in the case of Kubernetes), which changes at
every restart.

Because the hostname needs to be found in the KBD, we need a way to set
it to something stable. For now, I'm goign to experiment with an
environment variable.

Signed-off-by: Balthazar Rouberol <br@imap.cc>